backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
python-dotenv==1.0.1
python-multipart==0.0.18
pytesseract==0.3.13
pillow==11.1.0
google-generativeai==0.8.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
razorpay==1.4.2
pytest==8.3.4
httpx==0.28.1
